Rename: 엔티티 파일 이름 변경
- raw-data -> photo-booth-hidden로 변경

Related to #4
