+++ src/app.module.ts
@@ -10,7 +10,7 @@ import { User } from './user/entity/user.entity';
 import { AuthModule } from './auth/auth.module';
 import { Session } from './auth/entity/session.entity';
 import { PhotoBooth } from './photo-booth/entity/photo-booth.entity';
-import { HiddenPhotoBooth } from './photo-booth/entity/raw-data.entity';
+import { HiddenPhotoBooth } from './photo-booth/entity/photo-booth-hidden.entity';
 import {
   Hashtag,
   PhotoBoothBrand,
