+++ src/photo-booth/entity/photo-booth-hidden.entity.ts
