+++ src/photo-booth/photo-booth.module.ts
@@ -3,7 +3,7 @@ import { PhotoBoothService } from './photo-booth.service';
 import { PhotoBoothController } from './photo-booth.controller';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { PhotoBooth } from './entity/photo-booth.entity';
-import { HiddenPhotoBooth } from './entity/raw-data.entity';
+import { HiddenPhotoBooth } from './entity/photo-booth-hidden.entity';
 import { PhotoBoothRepository } from './repository/photo-booth.repository';
 import { HiddenBoothRepository } from './repository/photo-booth-hidden.repository';
 import {
